requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.2
django-allauth==0.43.0
django-appconf==1.0.4
django-compressor==2.4
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-jsonfield==1.4.1
django-libsass==0.8
djangorestframework==3.12.1
gunicorn==20.0.4
idna==2.10
libsass==0.20.1
Naked==0.1.31
numpy==1.19.4
oauthlib==3.1.0
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
rcssmin==1.0.6
requests==2.24.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
shellescape==3.8.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
